fixed error

+++ app/src/main/java/com/example/opsc_poe_part_2/DashboardActivity.kt
@@ -158,7 +158,7 @@ class DashboardActivity : AppCompatActivity() {
 
         // Retrieve data for each day of the week
         for (i in 0 until 7) {
-            val timeSpent = sharedPreferences.getInt("day_$i", 0).toDouble()
+            val timeSpent = sharedPreferences.getLong("day_$i", 0).toDouble()
             dataPoints.add(DataPoint(i.toDouble(), timeSpent))
         }
